feat: code to update object store (#6)
as it turns out, I can create a user however I am unable to
configure that user for access to bucket. Maybe this is
possible which is why I'm publishing the code, however the code
here doesn't get the job done.

Much of the docs for the Dell ECS appliance are hidden behind a
paywall making it difficult to understand what level of support they
provide to the boto3 client, and what is and isn't possible to
accomplish using the boto3 client.